Adding side panel to new zstack layers
